refactor(compositions): split validators into focused modules
Extract validator domain types and helper logic into dedicated modules while keeping the existing validators public API stable for composition flows.
