+++ docs/technical/AUDIT_NEXTJS_REACT_ACTION_PLAN.md
@@ -271,3 +271,4 @@ Pour chaque fichier (traiter par ordre métier / douleur) :
 | 2026-05-09 | — | Epic G (cleanup 3) : suppression des composants non utilisés `UserManagement`, `BackupRestore`, `GlobalStats`, `TeamManager`, `MatchHistoryManager`, `BurnoutRulesManager` et des modules associés |
 | 2026-05-09 | — | Epic G (lot 11) : extraction des dialogs `CreatePollDialog` / `ClosePollDialog` pour alléger `DiscordPollManager` |
 | 2026-05-09 | — | Epic G (lot 12) : extraction de `MentionMessageField` pour mutualiser l’édition + autocomplete des mentions dans `DiscordPollManager` |
+| 2026-05-09 | — | Epic G (lot 13) : `validators/index.ts` découpé en modules (`validators/types`, `team-utils`, `burnout-utils`, `paris-rules`) avec API publique conservée |
