+++ src/lib/compositions/validators/index.ts
@@ -1,339 +1,54 @@
 import { Player } from "@/types/team-management";
-import { ChampionshipType, Match } from "@/types";
-import { EquipeWithMatches } from "@/hooks/useTeamData";
 import { validateFFTTRules } from "@/lib/shared/fftt-utils";
-import { getMatchEpreuve, ID_EPREUVE_PARIS } from "@/lib/shared/epreuve-utils";
-import { getTeamsByType } from "@/lib/compositions/championship-utils";
+import {
+  type AssignmentValidationParams,
+  type AssignmentValidationResult,
+  DEFAULT_JOURNEE_RULE,
+  JOURNEE_CONCERNEE_PAR_REGLE,
+  type PhaseType,
+  type TeamCompositionValidationParams,
+  type TeamCompositionValidationResult,
+} from "@/lib/compositions/validators/types";
+import {
+  buildSimulatedPlayers,
+  didPlayerPlayJ1InLowerTeam,
+  extractTeamNumber,
+  findEquipeById,
+  getParisTeamStructure,
+  isParisChampionship,
+} from "@/lib/compositions/validators/team-utils";
+import {
+  calculateFutureBurnout,
+  calculateFutureBurnoutParis,
+} from "@/lib/compositions/validators/burnout-utils";
+import {
+  validateParisBurnoutByGroup,
+  validateParisGroupPointsOrder,
+} from "@/lib/compositions/validators/paris-rules";
 
-export type PhaseType = "aller" | "retour";
+export type {
+  AssignmentValidationParams,
+  AssignmentValidationResult,
+  CompositionMap,
+  PhaseType,
+  TeamCompositionValidationParams,
+  TeamCompositionValidationResult,
+} from "@/lib/compositions/validators/types";
 
-export interface CompositionMap {
-  [teamId: string]: string[];
-}
+export { JOURNEE_CONCERNEE_PAR_REGLE };
+export {
+  isMatchPlayed,
+  getPlayersFromMatch,
+  getMatchForTeamAndJournee,
+  extractTeamNumber,
+  getTeamNumberForPlayerJournee1,
+  didPlayerPlayJ1InLowerTeam,
+  isParisChampionship,
+  getParisTeamStructure,
+} from "@/lib/compositions/validators/team-utils";
+export { calculateFutureBurnout, calculateFutureBurnoutParis } from "@/lib/compositions/validators/burnout-utils";
+export { validateParisGroupPointsOrder, isPlayerBurnedParis, validateParisBurnoutByGroup } from "@/lib/compositions/validators/paris-rules";
 
-export interface AssignmentValidationParams {
-  playerId: string;
-  teamId: string;
-  players: Player[];
-  equipes: EquipeWithMatches[];
-  compositions: CompositionMap;
-  selectedPhase: PhaseType | null;
-  selectedJournee: number | null;
-  championshipType: ChampionshipType;
-  journeeRule?: number;
-  maxPlayersPerTeam?: number;
-}
-
-export interface AssignmentValidationResult {
-  canAssign: boolean;
-  reason?: string;
-  simulatedPlayers: Player[];
-  willBeBurned?: boolean; // Indique si le joueur sera brûlé en l'assignant à cette équipe
-  burnedTeamNumber?: number; // Numéro de l'équipe dans laquelle il sera brûlé
-}
-
-export interface TeamCompositionValidationParams {
-  teamId: string;
-  players: Player[];
-  equipes: EquipeWithMatches[];
-  compositions: CompositionMap;
-  selectedPhase: PhaseType | null;
-  selectedJournee: number | null;
-  championshipType: ChampionshipType;
-  journeeRule?: number;
-  maxPlayersPerTeam?: number;
-}
-
-export interface TeamCompositionValidationResult {
-  valid: boolean;
-  reason?: string;
-  offendingPlayerIds?: string[];
-}
-
-export const JOURNEE_CONCERNEE_PAR_REGLE = 2;
-
-const DEFAULT_JOURNEE_RULE = JOURNEE_CONCERNEE_PAR_REGLE;
-
-export const isMatchPlayed = (match: Match | null): boolean => {
-  if (!match) {
-    return false;
-  }
-
-  const hasPlayers =
-    match.joueursSQY &&
-    Array.isArray(match.joueursSQY) &&
-    match.joueursSQY.length > 0;
-
-  let hasValidScore = false;
-  const scoreValue = match.score;
-  const scoreType = typeof scoreValue;
-  const scoreIsString = scoreType === "string";
-  const scoreNotEmpty = scoreIsString && scoreValue !== "";
-  const scoreNotAVenir = scoreIsString && scoreValue !== "À VENIR";
-  const scoreNotZeroZero = scoreIsString && scoreValue !== "0-0";
-
-  if (
-    scoreValue &&
-    scoreIsString &&
-    scoreNotEmpty &&
-    scoreNotAVenir &&
-    scoreNotZeroZero
-  ) {
-    const scoreMatch = scoreValue.match(/^(\d+)-(\d+)$/);
-    if (scoreMatch !== null && scoreMatch.length === 3) {
-      const scoreA = parseInt(scoreMatch[1], 10);
-      const scoreB = parseInt(scoreMatch[2], 10);
-      hasValidScore = scoreA > 0 || scoreB > 0;
-    }
-  }
-
-  const validResults = ["VICTOIRE", "DEFAITE", "NUL", "ÉGALITÉ"];
-  const resultValue = match.result;
-  const resultType = typeof resultValue;
-  const resultIsString = resultType === "string";
-  const resultNotAVenir = resultIsString && resultValue !== "À VENIR";
-  const resultInValidList =
-    resultIsString &&
-    typeof resultValue === "string" &&
-    validResults.includes(resultValue.toUpperCase());
-  const hasValidResult =
-    hasValidScore &&
-    resultIsString &&
-    !!resultValue &&
-    resultNotAVenir &&
-    resultInValidList;
-
-  return hasPlayers || hasValidScore || hasValidResult;
-};
-
-export const getPlayersFromMatch = (
-  match: Match | null,
-  players: Player[]
-): Player[] => {
-  if (!match || !match.joueursSQY || !Array.isArray(match.joueursSQY)) {
-    return [];
-  }
-
-  return match.joueursSQY
-    .map((joueurSQY): Player | null => {
-      if (!joueurSQY.licence) return null;
-      return players.find((p) => p.license === joueurSQY.licence) || null;
-    })
-    .filter((p): p is Player => p !== null && p !== undefined);
-};
-
-export const getMatchForTeamAndJournee = (
-  equipe: EquipeWithMatches,
-  journee: number,
-  phase: PhaseType
-): Match | undefined => {
-  return equipe.matches.find(
-    (match) =>
-      match.journee === journee &&
-      match.phase?.toLowerCase() === phase.toLowerCase()
-  );
-};
-
-export const extractTeamNumber = (teamName: string): number => {
-  // Supporte les formats :
-  // - "SQY PING 3"
-  // - "SQY PING (3)"
-  // - "SQY PING 3 - Phase 1"
-  // - "SQY PING (3) - Phase 1"
-  // Cherche d'abord le format avec parenthèses après "SQY PING", puis n'importe quel nombre
-  if (teamName.includes("SQY PING")) {
-    const matchWithParentheses = teamName.match(/SQY PING\s*\((\d+)\)/i);
-    if (matchWithParentheses) {
-      return parseInt(matchWithParentheses[1], 10);
-    }
-    const match = teamName.match(/SQY PING\s*(\d+)/i);
-    if (match) {
-      return parseInt(match[1], 10);
-    }
-  }
-  // Fallback : chercher n'importe quel nombre
-  const match = teamName.match(/\b(\d+)\b/);
-  return match ? parseInt(match[1], 10) : 0;
-};
-
-const findEquipeById = (
-  equipes: EquipeWithMatches[],
-  teamId: string
-): EquipeWithMatches | undefined =>
-  equipes.find((equipe) => equipe.team.id === teamId);
-
-export const getTeamNumberForPlayerJournee1 = (
-  playerId: string,
-  phase: PhaseType,
-  players: Player[],
-  equipes: EquipeWithMatches[],
-  championshipType?: ChampionshipType
-): number | null => {
-  const player = players.find((p) => p.id === playerId);
-  if (!player) {
-    return null;
-  }
-
-  // Filtrer les équipes par type si championshipType est fourni
-  const equipesToCheck = championshipType
-    ? getTeamsByType(equipes)[championshipType]
-    : equipes;
-
-  for (const equipe of equipesToCheck) {
-    const matchJ1 = getMatchForTeamAndJournee(equipe, 1, phase);
-    if (matchJ1 && isMatchPlayed(matchJ1)) {
-      const playersFromMatch = getPlayersFromMatch(matchJ1, players);
-      const playerInMatch = playersFromMatch.find((p) => p.id === playerId);
-      if (playerInMatch) {
-        const teamNumber = extractTeamNumber(equipe.team.name);
-        return teamNumber > 0 ? teamNumber : null;
-      }
-    }
-  }
-
-  return null;
-};
-
-export const didPlayerPlayJ1InLowerTeam = (
-  playerId: string,
-  targetTeamNumber: number,
-  phase: PhaseType,
-  players: Player[],
-  equipes: EquipeWithMatches[],
-  championshipType?: ChampionshipType
-): boolean => {
-  const playerJ1TeamNumber = getTeamNumberForPlayerJournee1(
-    playerId,
-    phase,
-    players,
-    equipes,
-    championshipType
-  );
-
-  if (playerJ1TeamNumber === null) {
-    return false;
-  }
-
-  return playerJ1TeamNumber < targetTeamNumber;
-};
-
-const buildSimulatedPlayers = (
-  teamId: string,
-  player: Player,
-  players: Player[],
-  compositions: CompositionMap
-): Player[] => {
-  const currentTeamPlayers = compositions[teamId] || [];
-  return [
-    ...currentTeamPlayers
-      .filter((pid) => pid !== player.id)
-      .map((pid) => players.find((p) => p.id === pid))
-      .filter((p): p is Player => p !== undefined),
-    player,
-  ];
-};
-
-/**
- * Calcule le brûlage futur d'un joueur en simulant l'ajout d'un match dans une équipe
- * Utilise le même algorithme que team-matches-sync.ts :
- * - Crée une liste de tous les matchs triés par numéro d'équipe croissant
- * - Si le joueur a joué au moins 2 matchs, il est brûlé dans l'équipe du 2ème match
- * 
- * @param matchesByTeamByPhase - Map actuelle des matchs par équipe pour la phase (ex: {4: 1, 6: 2})
- * @param targetTeamNumber - Numéro de l'équipe où on ajoute le joueur
- * @returns Le numéro d'équipe de brûlage futur, ou null si le joueur ne sera pas brûlé (< 2 matchs)
- */
-export const calculateFutureBurnout = (
-  matchesByTeamByPhase: { [teamNumber: number]: number } | undefined,
-  targetTeamNumber: number,
-  _championshipType: ChampionshipType
-): number | null => {
-  void _championshipType;
-  // Créer une copie des matchs actuels
-  const futureMatchesByTeam = new Map<number, number>();
-  
-  if (matchesByTeamByPhase) {
-    for (const [teamNumber, matchCount] of Object.entries(matchesByTeamByPhase)) {
-      futureMatchesByTeam.set(parseInt(teamNumber, 10), matchCount);
-    }
-  }
-  
-  // Ajouter 1 match dans l'équipe cible
-  const currentCountInTargetTeam = futureMatchesByTeam.get(targetTeamNumber) || 0;
-  futureMatchesByTeam.set(targetTeamNumber, currentCountInTargetTeam + 1);
-  
-  // Créer une liste de tous les matchs triés par numéro d'équipe croissant
-  // (même algorithme que team-matches-sync.ts lignes 954-964)
-  const allMatches: number[] = [];
-  for (const [teamNumber, matchCount] of futureMatchesByTeam) {
-    // Ajouter le numéro d'équipe autant de fois qu'il y a de matchs
-    for (let i = 0; i < matchCount; i++) {
-      allMatches.push(teamNumber);
-    }
-  }
-  
-  // Trier par numéro d'équipe croissant
-  allMatches.sort((a, b) => a - b);
-  
-  // Si le joueur a joué au moins 2 matchs, il est brûlé dans l'équipe du 2ème match
-  if (allMatches.length >= 2) {
-    return allMatches[1]; // 2ème élément (index 1)
-  }
-  
-  return null; // Pas encore brûlé (< 2 matchs)
-};
-
-/**
- * Calcule le brûlage futur pour le championnat de Paris
- * Un joueur est brûlé s'il a joué 3 matchs ou plus dans UNE équipe de numéro inférieur
- * Retourne le numéro de l'équipe la plus basse (numéro le plus élevé) où il sera brûlé
- */
-export const calculateFutureBurnoutParis = (
-  matchesByTeamByPhase: { [teamNumber: number]: number } | undefined,
-  targetTeamNumber: number,
-  _championshipType: ChampionshipType
-): number | null => {
-  void _championshipType;
-  if (!matchesByTeamByPhase) {
-    return null;
-  }
-  
-  // Créer une copie des matchs actuels
-  const futureMatchesByTeam = new Map<number, number>();
-  for (const [teamNumber, matchCount] of Object.entries(matchesByTeamByPhase)) {
-    futureMatchesByTeam.set(parseInt(teamNumber, 10), matchCount);
-  }
-  
-  // Ajouter 1 match dans l'équipe cible
-  const currentCountInTargetTeam = futureMatchesByTeam.get(targetTeamNumber) || 0;
-  futureMatchesByTeam.set(targetTeamNumber, currentCountInTargetTeam + 1);
-  
-  // Pour le championnat de Paris, un joueur est brûlé s'il a 3 matchs ou plus dans UNE équipe de numéro inférieur
-  // On cherche la plus basse équipe (numéro le plus élevé) où il est brûlé
-  const teamNumbers = Array.from(futureMatchesByTeam.keys()).sort((a, b) => a - b);
-  let highestBurnedTeam: number | null = null;
-  
-  for (let i = 0; i < teamNumbers.length; i++) {
-    const currentTeamNumber = teamNumbers[i];
-    
-    // Vérifier s'il y a une équipe de numéro inférieur où le joueur a joué 3 fois ou plus
-    for (let j = 0; j < i; j++) {
-      const lowerTeamNumber = teamNumbers[j];
-      const matchCountInLowerTeam = futureMatchesByTeam.get(lowerTeamNumber) || 0;
-      
-      // Si le joueur a 3 matchs ou plus dans cette équipe inférieure, il est brûlé dans l'équipe actuelle
-      if (matchCountInLowerTeam >= 3) {
-        // Prendre la plus basse équipe où il est brûlé (numéro le plus élevé)
-        if (highestBurnedTeam === null || currentTeamNumber > highestBurnedTeam) {
-          highestBurnedTeam = currentTeamNumber;
-        }
-        break; // Une seule équipe inférieure suffit pour brûler
-      }
-    }
-  }
-  
-  return highestBurnedTeam;
-};
 
 export const canAssignPlayerToTeam = (
   params: AssignmentValidationParams
@@ -752,243 +467,3 @@ export const validateTeamCompositionState = (
 /**
  * Vérifie si une équipe fait partie du championnat de Paris
  */
-export const isParisChampionship = (
-  equipe: EquipeWithMatches
-): boolean => {
-  // Vérifier via idEpreuve
-  const match = equipe.matches[0];
-  if (match?.idEpreuve === ID_EPREUVE_PARIS) {
-    return true;
-  }
-  
-  // Vérifier via l'épreuve de l'équipe
-  const epreuve = getMatchEpreuve(match || {}, equipe.team);
-  if (epreuve === "championnat_paris") {
-    return true;
-  }
-  
-  // Fallback : vérifier le libellé de l'épreuve ou la division
-  const epreuveLibelle = equipe.team.epreuve?.toLowerCase() || "";
-  const division = equipe.team.division?.toLowerCase() || "";
-  
-  return (
-    epreuveLibelle.includes("paris idf") ||
-    epreuveLibelle.includes("excellence") ||
-    division.includes("excellence") ||
-    division.includes("paris idf")
-  );
-};
-
-/**
- * Détermine la structure d'une équipe du championnat de Paris selon sa division
- * @returns { groups: number, playersPerGroup: number, totalPlayers: number } ou null si non applicable
- */
-export const getParisTeamStructure = (
-  division: string
-): { groups: number; playersPerGroup: number; totalPlayers: number } | null => {
-  const divisionLower = division.toLowerCase();
-  
-  // 3 groupes (9 joueurs) : Excellence, Promo Excellence, Honneur
-  if (
-    divisionLower.includes("excellence") ||
-    divisionLower.includes("honneur")
-  ) {
-    return { groups: 3, playersPerGroup: 3, totalPlayers: 9 };
-  }
-  
-  // 2 groupes (6 joueurs) : 1ere Division ou 1ère Division
-  if (
-    divisionLower.includes("1ere division") ||
-    divisionLower.includes("1ère division") ||
-    divisionLower.includes("1ere div") ||
-    divisionLower.includes("1ère div")
-  ) {
-    return { groups: 2, playersPerGroup: 3, totalPlayers: 6 };
-  }
-  
-  // 1 groupe (3 joueurs) : 2eme Division ou 2ème Division
-  if (
-    divisionLower.includes("2eme division") ||
-    divisionLower.includes("2ème division") ||
-    divisionLower.includes("2eme div") ||
-    divisionLower.includes("2ème div")
-  ) {
-    return { groups: 1, playersPerGroup: 3, totalPlayers: 3 };
-  }
-  
-  return null;
-};
-
-/**
- * Valide l'ordre des points dans les groupes (Article 8 du championnat de Paris)
- * Chaque joueur du groupe 2 doit avoir des points ≤ groupe 1 et ≥ groupe 3
- * Permutation possible dans un même groupe
- */
-export const validateParisGroupPointsOrder = (
-  players: Player[],
-  structure: { groups: number; playersPerGroup: number; totalPlayers: number }
-): { valid: boolean; reason?: string } => {
-  if (structure.groups < 2) {
-    // Pas de validation nécessaire pour 1 seul groupe
-    return { valid: true };
-  }
-  
-  if (players.length !== structure.totalPlayers) {
-    // Pas encore complet, pas de validation
-    return { valid: true };
-  }
-  
-  // Trier les joueurs par points décroissants (les meilleurs en premier)
-  const sortedPlayers = [...players].sort((a, b) => {
-    const pointsA = a.points ?? 0;
-    const pointsB = b.points ?? 0;
-    return pointsB - pointsA; // Décroissant
-  });
-  
-  // Diviser les joueurs triés en groupes
-  const groups: Player[][] = [];
-  for (let i = 0; i < structure.groups; i++) {
-    const start = i * structure.playersPerGroup;
-    const end = start + structure.playersPerGroup;
-    groups.push(sortedPlayers.slice(start, end));
-  }
-  
-  // Pour chaque groupe, trouver min et max des points
-  const groupRanges = groups.map((group) => {
-    const points = group.map((p) => p.points || 0);
-    return {
-      min: Math.min(...points),
-      max: Math.max(...points),
-    };
-  });
-  
-  // Vérifier que groupe 2 respecte les contraintes
-  if (structure.groups >= 2) {
-    const group1Max = groupRanges[0].max;
-    // const group2Min = groupRanges[1].min;
-    // const group2Max = groupRanges[1].max;
-    
-    // Chaque joueur du groupe 2 doit avoir ≤ max du groupe 1
-    const group2Players = groups[1];
-    const invalidGroup2Players = group2Players.filter(
-      (p) => (p.points || 0) > group1Max
-    );
-    
-    if (invalidGroup2Players.length > 0) {
-      const playerNames = invalidGroup2Players
-        .map((p) => `${p.firstName} ${p.name} (${p.points || 0} pts)`)
-        .join(", ");
-      return {
-        valid: false,
-        reason: `Article 8 : Les joueurs du groupe 2 doivent avoir des points ≤ ${group1Max} pts (max du groupe 1). Joueurs non conformes : ${playerNames}`,
-      };
-    }
-    
-    // Si groupe 3 existe, chaque joueur du groupe 2 doit avoir ≥ min du groupe 3
-    if (structure.groups >= 3) {
-      const group3Min = groupRanges[2].min;
-      const invalidGroup2PlayersMin = group2Players.filter(
-        (p) => (p.points || 0) < group3Min
-      );
-      
-      if (invalidGroup2PlayersMin.length > 0) {
-        const playerNames = invalidGroup2PlayersMin
-          .map((p) => `${p.firstName} ${p.name} (${p.points || 0} pts)`)
-          .join(", ");
-        return {
-          valid: false,
-          reason: `Article 8 : Les joueurs du groupe 2 doivent avoir des points ≥ ${group3Min} pts (min du groupe 3). Joueurs non conformes : ${playerNames}`,
-        };
-      }
-    }
-  }
-  
-  return { valid: true };
-};
-
-/**
- * Vérifie si un joueur est brûlé selon les règles du championnat de Paris
- * Un joueur est brûlé s'il a joué 3 fois ou plus dans UNE équipe de numéro inférieur
- */
-export const isPlayerBurnedParis = (
-  player: Player,
-  teamNumber: number,
-  phase: PhaseType
-): boolean => {
-  // Pour le championnat de Paris, utiliser highestTeamNumberByPhaseParis
-  const burnedTeam = player.highestTeamNumberByPhaseParis?.[phase];
-  
-  if (burnedTeam === undefined || burnedTeam === null) {
-    return false;
-  }
-  
-  // Vérifier si le joueur a au moins 3 matchs dans UNE équipe inférieure
-  // En regardant les matchs par équipe pour le championnat de Paris
-  const matchesByTeam = player.matchesByTeamByPhaseParis?.[phase];
-  if (!matchesByTeam) {
-    return false;
-  }
-  
-  // Vérifier s'il y a une équipe de numéro inférieur où le joueur a joué 3 fois ou plus
-  for (const [teamNumStr, matchCount] of Object.entries(matchesByTeam)) {
-    const teamNum = parseInt(teamNumStr, 10);
-    if (teamNum < teamNumber && matchCount >= 3) {
-      return true; // Le joueur a 3+ matchs dans cette équipe inférieure
-    }
-  }
-  
-  return false;
-};
-
-/**
- * Valide le brûlage par groupe selon l'Article 12 du championnat de Paris
- * Pour équipes 2, 3, 4, etc., max 1 joueur brûlé par groupe de 3
- * Si 2 joueurs brûlés dans un groupe de 3, les 2 sont non qualifiés
- */
-export const validateParisBurnoutByGroup = (
-  players: Player[],
-  teamNumber: number,
-  structure: { groups: number; playersPerGroup: number; totalPlayers: number },
-  phase: PhaseType
-): { valid: boolean; reason?: string; offendingPlayerIds?: string[] } => {
-  // Cette règle ne s'applique que pour les équipes 2, 3, 4, etc. (pas l'équipe 1)
-  if (teamNumber <= 1) {
-    return { valid: true };
-  }
-  
-  if (players.length !== structure.totalPlayers) {
-    // Pas encore complet, pas de validation
-    return { valid: true };
-  }
-  
-  // Diviser les joueurs en groupes
-  const groups: Player[][] = [];
-  for (let i = 0; i < structure.groups; i++) {
-    const start = i * structure.playersPerGroup;
-    const end = start + structure.playersPerGroup;
-    groups.push(players.slice(start, end));
-  }
-  
-  // Vérifier chaque groupe de 3
-  for (let groupIndex = 0; groupIndex < groups.length; groupIndex++) {
-    const group = groups[groupIndex];
-    const burnedPlayers = group.filter((p) =>
-      isPlayerBurnedParis(p, teamNumber, phase)
-    );
-    
-    if (burnedPlayers.length > 1) {
-      const playerNames = burnedPlayers
-        .map((p) => `${p.firstName} ${p.name}`)
-        .join(", ");
-      return {
-        valid: false,
-        reason: `Article 12 : Groupe ${groupIndex + 1} : maximum 1 joueur brûlé par groupe de 3. ${burnedPlayers.length} joueurs brûlés détectés : ${playerNames}. Les 2 joueurs sont non qualifiés.`,
-        offendingPlayerIds: burnedPlayers.map((p) => p.id),
-      };
-    }
-  }
-  
-  return { valid: true };
-};
-
-
